crawler/requirements.txt
python-decouple
Requests
boto3
bs4
python-dotenv
responses